vacations filters

+++ backend/Logic/followerMYSQL.ts
@@ -1,11 +1,8 @@
-
 import { OkPacket } from "mysql";
 import dal_mysql from "../Utils/dal_mysql";
 
-
 const createFollowersTable = () => {
-    const SQLcommand = 
-    `
+  const SQLcommand = `
     CREATE TABLE IF NOT EXISTS followers (
         followerId INT NOT NULL AUTO_INCREMENT,
         userId INT NOT NULL,
@@ -21,30 +18,41 @@ const createFollowersTable = () => {
         ON UPDATE CASCADE
         );
         
-`
-    ;
-    const response = dal_mysql.execute(SQLcommand);
+`;
+  const response = dal_mysql.execute(SQLcommand);
 };
 
-const addLike = async (userId:number, vacationId:number) => {
-    const SQLcommand = `
+const addLike = async (userId: number, vacationId: number) => {
+  const SQLcommand = `
     INSERT INTO travel.followers (userId, vacationId) VALUES (${userId},${vacationId});
-    `
-    const response: OkPacket = await dal_mysql.execute(SQLcommand);
-    return response.insertId;      
-}
+    `;
+  const response: OkPacket = await dal_mysql.execute(SQLcommand);
+  return response.insertId;
+};
 
-const unLike = (vacationId:number) => {
-    const SQLcommand = `
-    DELETE FROM travel.followers WHERE vacationId = ${vacationId}`;
-    dal_mysql.execute(SQLcommand);
-    return true;
-}
-
-
+const unLike = (vacationId: number, userId: number) => {
+  const SQLcommand = `
+    DELETE FROM travel.followers WHERE vacationId = ${vacationId} AND userId=${userId}`;
+  dal_mysql.execute(SQLcommand);
+  return true;
+};
+const followersByVacationId = (vacationId: number) => {
+  const SQLcommand = `
+    SELECT * travel.followers WHERE vacationId = ${vacationId}`;
+  dal_mysql.execute(SQLcommand);
+  return true;
+};
+const followersByUserId = (userId: number) => {
+  const SQLcommand = `
+    SELECT * travel.followers WHERE userId = ${userId}`;
+  dal_mysql.execute(SQLcommand);
+  return true;
+};
 
 export default {
-    createFollowersTable,
-    addLike,
-    unLike
+  createFollowersTable,
+  addLike,
+  unLike,
+  followersByVacationId,
+  followersByUserId,
 };
